+++ src/lib/Serverless.ts
@@ -99,9 +99,29 @@ export class Serverless {
 		this.channel.show();
 
 		const serverlessCommand = `Running "serverless ${command} ${_.join(options, " ")}"`;
-		this.channel.appendLine(serverlessCommand);
+		this.channel.appendLine( serverlessCommand );
+
+		let timer: any = null;
+		let isTouched = false;
 
 		return new Promise( ( resolve, reject ) => {
+			// deploy되지 않은 어플리케이션을 호출할때 아무 이벤트도 받지 못하는 이슈때문에 사용 (20초 Timeout)
+			if ( /invoke --function/.test( command ) ) {
+				if ( timer ) clearTimeout(timer);
+				timer = setTimeout( () => {
+					if ( isTouched ) return;
+
+					window.showInputBox( {
+						prompt: `Confirm `,
+						placeHolder: 'Serverless 오류! 프로세스를 중지하려면 "YES"를 입력하세요.'
+					} ).then( value => {
+						if ( value !== 'YES' || isTouched ) return;
+						sls.kill( 'SIGHUP' );
+						reject();
+					} );
+				}, 20000 );
+			}
+
 			const sls = spawn( "node", _.concat(
 				[ `${nodeModulesPath}/serverless/bin/serverless` ],
 				_.split( command, " " ),
@@ -111,18 +131,23 @@ export class Serverless {
 				} );
 
 			sls.on( "error", err => {
+				isTouched = true;
 				reject( err );
 			} );
 
 			sls.stdout.on( "data", data => {
+				isTouched = true;
 				this.channel.append( data.toString() );
 			} );
 
 			sls.stderr.on( "data", data => {
+				isTouched = true;
 				this.channel.append( data.toString() );
 			} );
 
 			sls.on( "exit", code => {
+				isTouched = true;
+
 				if ( code !== 0 ) {
 					reject( new Error( `Command exited with ${code}` ) );
 				}