+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## 1.0.5
-- Invoke Function 옵션 추가
+- Invoke function 옵션 추가
+- Deploy service 하지 않은 상태에서 Invoke function 호출시 반응이 없을때 프로세스 중지 안내 추가
 - Remove 옵션 추가